[#1058] cluster code refactoring
