+++ collector/src/main/java/com/navercorp/pinpoint/collector/cluster/connection/ClusterConnectionManager.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2017 NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.navercorp.pinpoint.collector.cluster.connection;
+
+import java.net.InetSocketAddress;
+import java.net.SocketAddress;
+import java.util.List;
+
+/**
+ * @author minwoo.jung
+ */
+public interface ClusterConnectionManager {
+    void start();
+
+    void stop();
+
+    void connectPointIfAbsent(InetSocketAddress address);
+
+    public void disconnectPoint(SocketAddress address);
+
+    public List<SocketAddress> getConnectedAddressList();
+}
