+++ collector/src/main/java/com/navercorp/pinpoint/collector/cluster/zookeeper/ZookeeperClusterService.java
@@ -63,7 +63,7 @@ public class ZookeeperClusterService extends AbstractClusterService {
     private ZookeeperClient client;
 
     // WebClusterManager checks Zookeeper for the Web data, and manages collector -> web connections.
-    private ZookeeperWebClusterManager webClusterManager;
+    private ZookeeperClusterManager webClusterManager;
 
     // ProfilerClusterManager detects/manages profiler -> collector connections, and saves their information in Zookeeper.
     private ZookeeperProfilerClusterManager profilerClusterManager;
@@ -108,7 +108,7 @@ public class ZookeeperClusterService extends AbstractClusterService {
                     this.profilerClusterManager = new ZookeeperProfilerClusterManager(client, serverIdentifier, clusterPointRouter.getTargetClusterPointRepository());
                     this.profilerClusterManager.start();
 
-                    this.webClusterManager = new ZookeeperWebClusterManager(client, PINPOINT_WEB_CLUSTER_PATH, serverIdentifier, clusterConnectionManager);
+                    this.webClusterManager = new ZookeeperClusterManager(client, PINPOINT_WEB_CLUSTER_PATH, clusterConnectionManager);
                     this.webClusterManager.start();
 
                     this.serviceState.changeStateStarted();
@@ -187,7 +187,7 @@ public class ZookeeperClusterService extends AbstractClusterService {
         return profilerClusterManager;
     }
 
-    public ZookeeperWebClusterManager getWebClusterManager() {
+    public ZookeeperClusterManager getWebClusterManager() {
         return webClusterManager;
     }
