+++ web/src/main/java/com/navercorp/pinpoint/web/cluster/zookeeper/PushZnodeJob.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2017 NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.navercorp.pinpoint.web.cluster.zookeeper;
+
+import org.jboss.netty.util.TimerTask;
+
+/**
+ * @author minwoo.jung
+ */
+public interface PushZnodeJob extends TimerTask {
+    String getZNodePath();
+
+    byte[] getContents();
+
+    int getRetryInterval();
+}
