+++ web/src/main/java/com/navercorp/pinpoint/web/cluster/zookeeper/ZookeeperClusterDataManagerHelper.java
@@ -16,7 +16,6 @@
 
 package com.navercorp.pinpoint.web.cluster.zookeeper;
 
-import com.navercorp.pinpoint.web.cluster.zookeeper.ZookeeperClusterDataManager.PushWebClusterJob;
 import org.apache.zookeeper.CreateMode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -62,7 +61,7 @@ public class ZookeeperClusterDataManagerHelper {
         return Collections.emptyMap();
     }
 
-    String bindingPathAndZNode(String path, String zNodeName) {
+    public String bindingPathAndZNode(String path, String zNodeName) {
         StringBuilder fullPath = new StringBuilder();
 
         fullPath.append(path);
@@ -87,7 +86,7 @@ public class ZookeeperClusterDataManagerHelper {
         return null;
     }
 
-    boolean pushWebClusterResource(ZookeeperClient client, PushWebClusterJob job) {
+    public boolean pushZnode(ZookeeperClient client, PushZnodeJob job) {
         if (job == null) {
             return false;
         }
@@ -102,7 +101,7 @@ public class ZookeeperClusterDataManagerHelper {
 
             // ip:port zNode naming scheme
             String nodeName = client.createNode(zNodePath, contents, CreateMode.EPHEMERAL);
-            logger.info("Register Web Cluster Zookeeper UniqPath = {}.", zNodePath);
+            logger.info("Register Zookeeper node UniqPath = {}.", zNodePath);
             return true;
         } catch (Exception e) {
             logger.warn(e.getMessage(), e);
